rename hook file for consistency w/ name
